GenerateRulesets uses Utils.GetFilePath helper instead of raw paths
In particular this fixes case-sensitivity issues when opening RULES.TXT
on Linux. (See https://github.com/axx0/Civ2-clone/issues/142)

I manually tested this change in MGE with 'RULES.TXT' on a Linux system.
Prior to this commit, the game crashed on startup. With this change,
the game launches and I can launch a new game or load scenarios
without any crashes.