bugfix: load new-style SAV with 0-improvement cities without crashing
A city's improvements in JsonCityData is serialized as a boolean array,
and as an optimization all the trailing "falses" are cut off. A city with no
improvements is not written as an empty array - the json object doesn't
have the improvements key at all. (Example at the bottom of this commit
message.)

HydrateCity would try to iterate over the (null) array
cityData.Improvements and then crash with a null reference.

Example of such a save file:

> cat xe_ad180_new_style.sav.json | jq ".game.cities[] | .Name,.Improvements"

...
[
  "Rome",
  [
    false, true, false, false, true
  ]
]
[
  "London",
  [
    false, true, false, false, true
  ]
]
[
  "Hamburg",
  null
]
...