Store CitiesBuiltSoFar in GameData and use that for city name offset, rather than relying on cities built according to History
The change to use History was wrong for reasons already pointed out;
that it doesn't account for cities razed or captured by other civs.

This is similar to the original approach used in 5807d4dc where city
count was stored in the Civilization class.

This is not a complete implementation. Currently city name is chosen
correctly when loading classic SAV files but not when saving/loading
with the new JSON format.