Store CitiesBuiltSoFar in GameData and use that for city name offset, rather than relying on cities built according to History
The change to use History was wrong for reasons already pointed out;
that it doesn't account for cities razed or captured by other civs.

This is similar to the original approach used in 5807d4dc where city
count was stored in the Civilization class.

I manually tested:
* Starting a new game (as Romans) and verifying that the first two
  cities founded are Rome and Veii
* Loading an old-style SCN (MARS.SCN), where each player starts with
  one city; second city founded is named correctly.
* Loading an old style SAV with lost cities to verify offsets are loaded
  correctly.

Further work still needed for new-format JSON SAV files,
where CitiesBuiltSoFar reverts back to 0 on load.